ingestion_service/requirements.txt
opencv-python
boto3
python-dotenv
botocore
